Now able to render sphere! -Added code to check for shpere intersection. -Added viewpoint and rays for each pixel. -Improved code readability.

+++ Ray Tracer/Ray Tracer.h
@@ -12,6 +12,6 @@ using namespace std;
 
 void WritePPM(const vector<vector<Vec3>>& image, const string& filename);
 void GetImage(vector<vector<Vec3>>& image, const int width, const int height);
-Vec3 GetColor(const int i, const int j);
+Vec3 GetColor(const float i, const float j);
 
 bool hits_sphere(const Sphere sphere, const Ray ray);
